Conflicts:
	modules/frontend-web-dossier-bvh/.settings/org.eclipse.buildship.core.prefs
	modules/frontend-web-payment2/.settings/org.eclipse.buildship.core.prefs
	modules/swagger-generate/src/main/java/org/swagger/model/verified/Verified.java

+++ modules/backend-api-rest/src/main/java/org/opencps/api/serviceprocess/model/ServiceProcessDetailModel.java
@@ -79,7 +79,7 @@ public class ServiceProcessDetailModel {
     protected String processNo;
     protected String processName;
     protected String description;
-    protected Integer durationCount;
+    protected Double durationCount;
     protected Integer durationUnit;
     protected Integer counter;
     protected String generateDossierNo;
@@ -252,7 +252,7 @@ public class ServiceProcessDetailModel {
      *     {@link Integer }
      *     
      */
-    public Integer getDurationCount() {
+    public Double getDurationCount() {
         return durationCount;
     }
 
@@ -264,7 +264,7 @@ public class ServiceProcessDetailModel {
      *     {@link Integer }
      *     
      */
-    public void setDurationCount(Integer value) {
+    public void setDurationCount(Double value) {
         this.durationCount = value;
     }
